test: setup function  refactored e2e block building test
